checkpoint - should scroll the entities
Header, Main and Footer are in place. Header and footer are fixed top and bottom. The entitles are scrolled in the main portion.

When the header goes into a locked mode, it's jumping into the first row - that needs to be addressed.
The footer doesn't seem to leave enough room for iPhoneX - the activity bar seems to cover the buttons to save/delete
I need to teach the main window how to increase/decrease the number of entries being shown so they fit.